telegram_bot/requirements.txt
sounddevice>=0.4.6
numpy>=1.21.0
openai-whisper>=20231117
scipy>=1.7.0
requests>=2.25.0
pyttsx3>=2.90 